Implement function to accept/reject event and checkin/cancel enrollment: resolve #105 (#106)

+++ src/main/java/io/lcalmsky/app/event/application/EventService.java
@@ -45,8 +45,26 @@ public class EventService {
 
     public void leave(Event event, Account account) {
         Enrollment enrollment = enrollmentRepository.findByEventAndAccount(event, account);
-        event.removeEnrollment(enrollment);
-        enrollmentRepository.delete(enrollment);
-        event.acceptNextIfAvailable();
+        if (!enrollment.isAttended()) {
+            event.removeEnrollment(enrollment);
+            enrollmentRepository.delete(enrollment);
+            event.acceptNextIfAvailable();
+        }
+    }
+
+    public void acceptEnrollment(Event event, Enrollment enrollment) {
+        event.accept(enrollment);
+    }
+
+    public void rejectEnrollment(Event event, Enrollment enrollment) {
+        event.reject(enrollment);
+    }
+
+    public void checkInEnrollment(Event event, Enrollment enrollment) {
+        enrollment.attend();
+    }
+
+    public void cancelCheckinEnrollment(Event event, Enrollment enrollment) {
+        enrollment.absent();
     }
 }
